+++ aiida/backends/tests/tools/importexport/test_specific_import.py
@@ -1,0 +1,162 @@
+# -*- coding: utf-8 -*-
+###########################################################################
+# Copyright (c), The AiiDA team. All rights reserved.                     #
+# This file is part of the AiiDA code.                                    #
+#                                                                         #
+# The code is hosted on GitHub at https://github.com/aiidateam/aiida_core #
+# For further information on the license, see the LICENSE.txt file        #
+# For further information please visit http://www.aiida.net               #
+###########################################################################
+"""Tests for the export and import routines"""
+from __future__ import division
+from __future__ import print_function
+from __future__ import absolute_import
+from __future__ import with_statement
+
+import tempfile
+
+from aiida import orm
+from aiida.backends.testbase import AiidaTestCase
+from aiida.tools.importexport import import_data, export
+
+
+class TestSpecificImport(AiidaTestCase):
+    """Test specific ex-/import cases"""
+
+    def setUp(self):
+        super(TestSpecificImport, self).setUp()
+        self.reset_database()
+
+    def tearDown(self):
+        self.reset_database()
+
+    def test_simple_import(self):
+        """
+        This is a very simple test which checks that an export file with nodes
+        that are not associated to a computer is imported correctly. In Django
+        when such nodes are exported, there is an empty set for computers
+        in the export file. In SQLA there is such a set only when a computer is
+        associated with the exported nodes. When an empty computer set is
+        found at the export file (when imported to an SQLA profile), the SQLA
+        import code used to crash. This test demonstrates this problem.
+        """
+        parameters = orm.Dict(
+            dict={
+                'Pr': {
+                    'cutoff': 50.0,
+                    'pseudo_type': 'Wentzcovitch',
+                    'dual': 8,
+                    'cutoff_units': 'Ry'
+                },
+                'Ru': {
+                    'cutoff': 40.0,
+                    'pseudo_type': 'SG15',
+                    'dual': 4,
+                    'cutoff_units': 'Ry'
+                },
+            }).store()
+
+        with tempfile.NamedTemporaryFile() as handle:
+            nodes = [parameters]
+            export(nodes, outfile=handle.name, overwrite=True, silent=True)
+
+            # Check that we have the expected number of nodes in the database
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), len(nodes))
+
+            # Clean the database and verify there are no nodes left
+            self.clean_db()
+            self.create_user()
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), 0)
+
+            # After importing we should have the original number of nodes again
+            import_data(handle.name, silent=True)
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), len(nodes))
+
+    def test_cycle_structure_data(self):
+        """
+        Create an export with some orm.CalculationNode and Data nodes and import it after having
+        cleaned the database. Verify that the nodes and their attributes are restored
+        properly after importing the created export archive
+        """
+        from aiida.common.links import LinkType
+
+        test_label = 'Test structure'
+        test_cell = [[8.34, 0.0, 0.0], [0.298041701839357, 8.53479766274308, 0.0],
+                     [0.842650688117053, 0.47118495164127, 10.6965192730702]]
+        test_kinds = [{
+            'symbols': [u'Fe'],
+            'weights': [1.0],
+            'mass': 55.845,
+            'name': u'Fe'
+        }, {
+            'symbols': [u'S'],
+            'weights': [1.0],
+            'mass': 32.065,
+            'name': u'S'
+        }]
+
+        structure = orm.StructureData(cell=test_cell)
+        structure.append_atom(symbols=['Fe'], position=[0, 0, 0])
+        structure.append_atom(symbols=['S'], position=[2, 2, 2])
+        structure.label = test_label
+        structure.store()
+
+        parent_process = orm.CalculationNode()
+        parent_process.set_attribute('key', 'value')
+        parent_process.store()
+        child_calculation = orm.CalculationNode()
+        child_calculation.set_attribute('key', 'value')
+        remote_folder = orm.RemoteData(computer=self.computer, remote_path='/').store()
+
+        remote_folder.add_incoming(parent_process, link_type=LinkType.CREATE, link_label='link')
+        child_calculation.add_incoming(remote_folder, link_type=LinkType.INPUT_CALC, link_label='link')
+        child_calculation.store()
+        structure.add_incoming(child_calculation, link_type=LinkType.CREATE, link_label='link')
+
+        with tempfile.NamedTemporaryFile() as handle:
+
+            nodes = [structure, child_calculation, parent_process, remote_folder]
+            export(nodes, outfile=handle.name, overwrite=True, silent=True)
+
+            # Check that we have the expected number of nodes in the database
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), len(nodes))
+
+            # Clean the database and verify there are no nodes left
+            self.clean_db()
+            self.create_user()
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), 0)
+
+            # After importing we should have the original number of nodes again
+            import_data(handle.name, silent=True)
+            self.assertEqual(orm.QueryBuilder().append(orm.Node).count(), len(nodes))
+
+            # Verify that orm.CalculationNodes have non-empty attribute dictionaries
+            builder = orm.QueryBuilder().append(orm.CalculationNode)
+            for [calculation] in builder.iterall():
+                self.assertIsInstance(calculation.attributes, dict)
+                self.assertNotEqual(len(calculation.attributes), 0)
+
+            # Verify that the structure data maintained its label, cell and kinds
+            builder = orm.QueryBuilder().append(orm.StructureData)
+            for [structure] in builder.iterall():
+                self.assertEqual(structure.label, test_label)
+                self.assertEqual(structure.cell, test_cell)
+
+            builder = orm.QueryBuilder().append(orm.StructureData, project=['attributes.kinds'])
+            for [kinds] in builder.iterall():
+                self.assertEqual(len(kinds), 2)
+                for kind in kinds:
+                    self.assertIn(kind, test_kinds)
+
+            # Check that there is a StructureData that is an output of a orm.CalculationNode
+            builder = orm.QueryBuilder()
+            builder.append(orm.CalculationNode, project=['uuid'], tag='calculation')
+            builder.append(orm.StructureData, with_incoming='calculation')
+            self.assertGreater(len(builder.all()), 0)
+
+            # Check that there is a RemoteData that is a child and parent of a orm.CalculationNode
+            builder = orm.QueryBuilder()
+            builder.append(orm.CalculationNode, tag='parent')
+            builder.append(orm.RemoteData, project=['uuid'], with_incoming='parent', tag='remote')
+            builder.append(orm.CalculationNode, with_incoming='remote')
+            self.assertGreater(len(builder.all()), 0)
